:art:Implemented Protection for bootcamp ownership

+++ controllers/bootcamps.js
@@ -34,21 +34,21 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route     POST /api/v1/bootcamps
 // @access    Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
-  // // Add user to req,body
-  // req.body.user = req.user.id;
+  // Add user to req.body
+  req.body.user = req.user.id;
 
-  // // Check for published bootcamp
-  // const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+  // Check for published bootcamp
+  const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
 
-  // // If the user is not an admin, they can only add one bootcamp
-  // if (publishedBootcamp && req.user.role !== 'admin') {
-  //   return next(
-  //     new ErrorResponse(
-  //       `The user with ID ${req.user.id} has already published a bootcamp`,
-  //       400
-  //     )
-  //   );
-  // }
+  // If the user is not an admin, they can only add one bootcamp
+  if (publishedBootcamp && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `The user with ID ${req.user.id} has already published a bootcamp`,
+        400
+      )
+    );
+  }
 
   const bootcamp = await Bootcamp.create(req.body);
 
@@ -62,28 +62,31 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route     PUT /api/v1/bootcamps/:id
 // @access    Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  // let bootcamp = await Bootcamp.findById(req.params.id);
+  let bootcamp = await Bootcamp.findById(req.params.id);
 
-  // if (!bootcamp) {
-  //   return next(
-  //     new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
-  //   );
-  // }
+
+  if (!bootcamp) {
+    return next(
+      new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+    );
+  }
 
   // // Make sure user is bootcamp owner
-  // if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
-  //   return next(
-  //     new ErrorResponse(
-  //       `User ${req.params.id} is not authorized to update this bootcamp`,
-  //       401
-  //     )
-  //   );
-  // }
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.params.id} is not authorized to update this bootcamp`,
+        401
+      )
+    );
+  }
 
-  bootcamp = await Bootcamp.findOneAndUpdate(req.params.id, req.body, {
-    new: true,
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id , req.body , {
+    new : true,
     runValidators: true
-  });
+  })
+
+
 
   res.status(200).json({ success: true, data: bootcamp });
 });
@@ -101,14 +104,14 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
   }
 
   // Make sure user is bootcamp owner
-  // if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
-  //   return next(
-  //     new ErrorResponse(
-  //       `User ${req.params.id} is not authorized to delete this bootcamp`,
-  //       401
-  //     )
-  //   );
-  // }
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.params.id} is not authorized to delete this bootcamp`,
+        401
+      )
+    );
+  }
 
   bootcamp.remove(); // This activates the remove middleware in Bootcamp model ; findByIdAndDelete won't
 
@@ -155,14 +158,14 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   }
 
   // Make sure user is bootcamp owner
-  // if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
-  //   return next(
-  //     new ErrorResponse(
-  //       `User ${req.params.id} is not authorized to update this bootcamp`,
-  //       401
-  //     )
-  //   );
-  // }
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.params.id} is not authorized to update this bootcamp`,
+        401
+      )
+    );
+  }
 
   if (!req.files) {
     return next(new ErrorResponse(`Please upload a file`, 400));
